Make RoutingHttpProxy class to expose
Before, I was just exposing a function that would take a stream to route.
That function also took the get-target-stream callback. But then I also
wanted to separate out the error reporting strategy. That would mean
taking two callbacks or something, which is getting messy. Instead, you
construct a RoutingHttpProxy, giving it the get-target-stream callback,
and handle error events that it emits.